pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = "metabsae embedder"
authors = ["shahnoor77"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.1.0"
httpx = "^0.26.0"
python-multipart = "^0.0.6"
pyjwt = "^2.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
black = "^24.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"